% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinmap.R
\name{add_facility_unit_variable}
\alias{add_facility_unit_variable}
\title{Split facility-unit information from a input csv filename}
\usage{
add_facility_unit_variable(input_csv)
}
\arguments{
\item{input_csv}{Filename of a csv file with ID variable
formatted as Facility-Unit}
}
\value{
A data frame that adds Facility and Utility variable to the CSV file
}
\description{
\code{add_facility_unit_variable} adds facility-unit information
from a input csv filename.
}
