% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinmap.R
\name{combine_inmap_output}
\alias{combine_inmap_output}
\title{Combine zipcode-linked inMAP output into single data object and link with spatial
data for plotting.}
\usage{
combine_inmap_output(path.out, pattern)
}
\arguments{
\item{path.out}{Directory that houses InMAP output csv files}

\item{pattern}{A text or regex pattern common to all InMAP output you wish to be joined}

\item{zcta_shapefile}{A ZCTA shapefile}
}
\value{
A data frame of average PM2.5 values at the ZIP code level.
}
\description{
\code{combine_inmap_output} uses spatial linkages of an inMAP outputs to combine output
multiple runs. Employs the sf package to link with spatial data if it is installed.
}
