% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinmap.R
\name{create_input_facility_all_but_one}
\alias{create_input_facility_all_but_one}
\title{Saves csv files taking one unit at a time}
\usage{
create_input_facility_all_but_one(input_csv, path = basename(input_csv))
}
\arguments{
\item{input_csv}{Filename of a csv file to be read and converted to a shapefile format.}

\item{path}{Folder where the csv files are saved}
}
\value{
This function does not return anything and is used for its side-effects.
}
\description{
\code{create_input_facility_all_but_one} saves one file taking all facilities but one as csv files
to a specified directory.
}
