% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinmap.R
\name{create_input_shapefile}
\alias{create_input_shapefile}
\title{Convert csv file to inMAP-compatible shapefile}
\usage{
create_input_shapefile(input_csv, path)
}
\arguments{
\item{input_csv}{Filename of a csv file to be read and converted to a shapefile format.
Expected variables are:
#' \enumerate{
  \item Latitude, Longitude (in NAD83 CRS)
  \item SOx, CO2, NOx (in tons per year)
  \item Height, Diam(in meters)
  \item Velocity (in meters per second)
  \item Temp (in Kelvin)
}}

\item{path}{Folder where the shaper is going to be saved
(in a subfolder, as "emis/ptegu.shp")}
}
\value{
This function does not return anything and is used for its side-effects.
}
\description{
\code{create_input_shapefile} saves an inMAP-compatible shapefile to a specified directory.
}
