% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinmap.R
\name{plot_inmap}
\alias{plot_inmap}
\title{Using zipcode-linked inMAP output from \code{combine_inmap_output}, plot change in InMAP impacts
at zip code level compared to base year.}
\usage{
plot_inmap(read_inmap_d,
  zcta_shapefile = "~/shared_space/ci3_nsaph/software/inmap/zcta/cb_2015_us_zcta510_500k.shp",
  legend_lims = c(-5, 5), path.plot = "InMAP_plots", cores = 1)
}
\arguments{
\item{read_inmap_d}{Directory that houses InMAP output csv files}

\item{legend_lims}{Legend limits for zip code fill}

\item{path.plot}{Output directory for plots. If it does not exist, it will be created}

\item{cores}{Cores available to create plots. Defaults to one}
}
\value{
A list of ggplot objects.
}
\description{
\code{combine_inmap_output} uses spatial linkages of an inMAP outputs to combine output
multiple runs. Requires the sf, ggplot2, and parallel packages.
}
