% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinmap.R
\docType{package}
\name{rinmap}
\alias{rinmap}
\alias{rinmap-package}
\title{rinmap: An R interface to inMAP.}
\description{
The rinmap package provides an R interface to inMAP
(Intervention Model for Air Pollution, \url{http://spatialmodel.com/inmap/}).
}
