% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinmap.R
\name{run_inmap}
\alias{run_inmap}
\title{Call inMAP and save ZIP code level average PM2.5}
\usage{
run_inmap(input_csv,
  zcta_shapefile = "~/shared_space/ci3_nsaph/software/inmap/zcta/cb_2015_us_zcta510_500k.shp",
  crosswalk_csv = "~/shared_space/ci3_nsaph/software/inmap/crosswalk/Zip_to_ZCTA_crosswalk_2015_JSI.csv")
}
\arguments{
\item{input_csv}{Filename of a csv file to be read and converted to a shapefile format.}

\item{zcta_shapefile}{A ZCTA shapefile (default value is hard-coded cluster path)}

\item{crosswalk_csv}{A crosswalk csv file to convert ZCTA to ZIP (default value is hard-coded cluster path)}
}
\value{
A data frame with ZIP code level average PM2.5 values.
}
\description{
\code{run_inmap} calls the inMAP executable.  Paths are hard-coded in file \code{inmap.toml}.
}
