% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinmap.R
\name{setup_files_inmap_run}
\alias{setup_files_inmap_run}
\title{Copy inMAP setup files to the run folder}
\usage{
setup_files_inmap_run(path)
}
\arguments{
\item{path}{Filepath where \code{inmap.toml} and \code{run.sh} are copied.}
}
\value{
This function does not return anything and is used for its side-effects.
}
\description{
\code{setup_files_inmap_run} copies \code{inmap.toml} and \code{run.sh} to a specified
folder (typically the folder where inMAP is going to be run).
}
