% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinmap.R
\name{zip_code_linkage}
\alias{zip_code_linkage}
\title{Link inMAP output at the ZIP code level.}
\usage{
zip_code_linkage(output_shapefile, zcta_shapefile, crosswalk_csv)
}
\arguments{
\item{output_shapefile}{An inMAP output shapefile}

\item{zcta_shapefile}{A ZCTA shapefile}

\item{crosswalk_csv}{A crosswalk csv file to convert ZCTA to ZIP}
}
\value{
A data frame of average PM2.5 values at the ZIP code level.
}
\description{
\code{zip_code_linkage} performs spatial linkage of an inMAP output shapefile
at the ZIP code level.
}
